pyproject.toml
[project]
name = "teledrop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "argon2-cffi>=23.1.0",
    "fastapi[standard]>=0.115.8",
    "pydantic-settings>=2.7.1",
    "pyjwt>=2.10.1",
    "sqlmodel>=0.0.22",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.0",
    "ruff>=0.11.12",
]
